modularized and optimized code
