+++ core/app/[locale]/(default)/test-ride-waiver/page.tsx
@@ -30,9 +30,9 @@ export default async function TestRideWaiverPage({ params }: Props) {
   return (
     <>
       <PageHeader 
-        backgroundImage={Streamable.from(async () => Promise.resolve(backgroundImage))}
+        backgroundImage={backgroundImage}
         locale={locale}
-        title={Streamable.from(async () => Promise.resolve('Test Ride Waiver'))}
+        title="Test Ride Waiver"
       />
       <TermsPage
         content={{
