+++ core/components/terms-page/price-match-content.tsx
@@ -2,18 +2,28 @@
 
 import { ButtonLink } from '@/vibes/soul/primitives/button-link';
 
-export function PriceMatchContent() {
+interface Props {
+  buttonText?: string;
+  href?: string;
+  className?: string;
+}
+
+export function PriceMatchContent({ 
+  buttonText = "Visit Shop", 
+  href = "/", 
+  className = "bg-pink-500 hover:bg-pink-600 text-white" 
+}: Props = {}) {
   return (
     <div className="space-y-8">
       <div>
         <ButtonLink 
-          className="bg-pink-500 hover:bg-pink-600 text-white"
-          href="/"
+          className={className}
+          href={href}
           shape="pill"
           size="small"
           variant="primary"
         >
-          Visit Shop
+          {buttonText}
         </ButtonLink>
       </div>
     </div>
