+++ core/lib/config/company-info.ts
@@ -1,0 +1,11 @@
+export const COMPANY_INFO = {
+  copyrightAgent: {
+    company: 'RYDDO LLC.',
+    address: '1200 S Figueroa St #3214',
+    city: 'Los Angeles',
+    state: 'CA',
+    zip: '90015',
+    attention: 'c/o Warwick Hunt',
+    email: 'warwick@ryddo.com'
+  }
+};
