+++ core/lib/data/terms-conditions/index.ts
@@ -1,0 +1,7 @@
+export * from './interfaces';
+export * from './navigation';
+export * from './terms-of-use';
+export * from './shipping-returns';
+export * from './price-match';
+export * from './privacy-policy';
+export * from './test-ride-waiver';
