+++ core/lib/data/terms-conditions/price-match.ts
@@ -1,0 +1,16 @@
+import type { TermsSection } from './interfaces';
+
+// Price Match Guarantee page content
+export const priceMatchContent: TermsSection = {
+  id: 'price-match',
+  title: 'We Price Match<span class="text-pink-600">!</span>',
+  pageTitle: 'Price Match Guarantee',
+  description: `We will price match any written or online offer from a U.S. based authorized reseller for the exact same product, tax excluded. The price match must be requested prior to purchase. Ebay and other unauthorized resellers excluded. Please contact us for more details. The customer may submit a request online using the "CLICK TO PRICE MATCH" button on the product pages.`,
+  accordionItems: [
+    {
+      id: 'main-content',
+      title: '',
+      content: `<h2 class="text-3xl font-bold text-gray-900 mb-6 text-center">We Price Match! <span class="text-pink-600">!</span></h2>We will price match any written or online offer from a U.S. based authorized reseller for the exact same product, tax excluded. The price match must be requested prior to purchase. Ebay and other unauthorized resellers excluded.<br /><br />Please contact us for more details. The customer may submit a request online using the "CLICK TO PRICE MATCH" button on the product pages.`
+    }
+  ]
+};
