+++ core/lib/data/terms-conditions/privacy-policy.ts
@@ -1,0 +1,288 @@
+import type { TermsSection } from './interfaces';
+
+// Privacy Policy page content
+export const privacyPolicyContent: TermsSection = {
+  id: 'privacy-policy',
+  title: 'Privacy Policy',
+  pageTitle: 'Privacy Policy',
+  description: `RYDDO LLC. and its affiliates ("RYDDO" or "we", "our", "us") respects the privacy of our consumers. This Privacy Policy ("Privacy Policy") explains the type of information we collect from individuals that visit our website(s) (each, a "Site") where this Privacy Policy is posted or otherwise connect with us offline such as via telephone ("Offline Services", and collectively with all Sites, the "Services" or individually, a "Service"), and how we store, use, disclose and protect such information. By using or accessing our Services, you agree to the collection, storage, use and disclosure of your information as described in this Privacy Policy.`,
+  accordionItems: [
+    {
+      id: 'what-information-we-collect',
+      title: 'WHAT INFORMATION WE COLLECT',
+      content: '',
+      subsections: [
+        {
+          title: 'Information you give us.',
+          content: 'We collect and maintain information that you may provide to us through use of our Services, including shipping address, billing address, credit card information, telephone number(s), email address(es) and purchase/return/exchange data. For example, when you set up your account and/or make online purchases, you will be asked to provide information necessary to facilitate the transaction such as contact and payment information; when you enter our contests, sweepstakes or other promotions, you may be asked for information necessary to fulfill the promotion; when you post comments or other content in publicly available forums on our Services (e.g., Product reviews, Q&A forum), anything you post will be available to us as well as visitors to our Services; when you post information about us on third party social media platforms (each, a "Social Media Service"), we may collect such information through an integration with the Social Media Service provider; when you contact our customer service representatives, we collect additional information you choose to provide our representatives.'
+        },
+        {
+          title: 'Information automatically collected.',
+          content: 'We may use technology to automatically collect other information about you when you visit our Services or view our advertisements. See Cookies and Interest Based Advertising, below. This information helps our online Services work correctly and supports our customer marketing and analytics efforts. This information includes, and is not limited to, device type, browser information, operating system information, device IP, device information, IP address, mobile device identifier, software version, phone model, phone operating system, carrier information, geo-location, the date, time, length of stay, referring website and specific pages accessed and other actions you take on our online Service.'
+        },
+        {
+          title: 'Information collected from other sources.',
+          content: 'We receive information about you from other sources to help us correct or supplement our records, improve the quality of personalization of our Services to you and prevent or detect fraud. We may use this information consistent with this Privacy Policy.'
+        },
+        {
+          title: 'Information collected through Social Media Services.',
+          content: 'You may be able to access our Services through Social Media Services. If you log on to our Services through a Social Media Service, you give us permission to access, store and use any information that you permit the applicable Social Media Service to share with us in accordance with your privacy settings with the applicable Social Media Service. When using Social Media Services with our Services, the Social Media Service provider will be able to collect information from you. We may also place tags and other data collection mechanisms on our Services from Social Media Services through which they can collect information about the usage of our Services and their services across multiple devices (such as your smartphone and your computer) for use by both us and such Social Media Service. We are not responsible or liable for the privacy practices or content of the applicable Social Media Service. We encourage you to review, and if necessary, adjust, your privacy settings with the applicable Social Media Service before logging in to our online Services using the applicable Social Media Service. Information that we receive from Social Media Services is stored and use by us in accordance with this Privacy Policy.'
+        }
+      ]
+    },
+    {
+      id: 'how-we-use',
+      title: 'HOW WE USE YOUR INFORMATION',
+      content: '',
+      description: 'We use, process and disclose your information we collect for our business purposes, including:',
+      subsections: [
+        {
+          title: 'Responding to your questions and requests.',
+          content: 'We use your information to provide you with transactional assistance such as fulfilling your order and ensuring proper delivery of your products, as well as administering promotions such sweepstakes and contests and responding to reviews, comments or other feedback you provide.'
+        },
+        {
+          title: 'Communicating about your account.',
+          content: 'We may contact you to tell you about changes to this Privacy Policy, our Terms of Use or changes to our Services, products and programs. We may also tell you about issues with your orders or if there is a product recall.'
+        },
+        {
+          title: 'Improving our products and services.',
+          content: 'We may use your information to make improvements to our Sites, including to personalize our service offerings to meet your needs.'
+        },
+        {
+          title: 'Reviewing Site trends and customer interests.',
+          content: 'We may use your information to customize your experience with us. We may collect information about your activities on and interactions with various devices and link that information. Through such cross-device linking, we can provide you with a consistent experience across the devices you use. We may also combine information we get from you with information about you we have received from third parties or publicly available sources to assess trends and interests.'
+        },
+        {
+          title: 'Improving marketing communications.',
+          content: 'We may send you communications about promotions or offers via regular mail, email or other electronic channels, including ads on Social Media Services. We may use information, including information collected across different online services and collected from the various devices you may use, to deliver marketing communications based on your interests.'
+        },
+        {
+          title: 'Promote our products or services.',
+          content: 'If you complete a product review or other feedback or provide any image of yourself, we may use your submission for commercial purposes to promote our products and services.'
+        },
+        {
+          title: 'Other uses we may disclose',
+          content: 'We also may use your information with your consent, and as otherwise permitted or required by law.'
+        }
+      ]
+    },
+    {
+      id: 'what-information-we-disclose',
+      title: 'WHAT INFORMATION WE DISCLOSE',
+      content: '',
+      description: 'RYDDO does not rent, sell, trade or disclose your information to unrelated third parties, except as set forth in this Privacy Policy. We may share your information in limited circumstances, such as to conduct our business, when legally required or with your consent, including:',
+      subsections: [
+        {
+          title: 'Affiliates.',
+          content: 'We may share your information with our affiliates for business, operational, promotional, marketing, or other purposes consistent with this Privacy Policy.'
+        },
+        {
+          title: 'Third-Party Service Providers.',
+          content: 'We may disclose the information collected about you to our third-party contractors and partners relating to their performing services or other activities in support of our Services and/or our businesses, or their completing or confirming on our behalf a transaction or series of transactions that you conduct with us. Examples include third parties which may fulfill orders, deliver packages, send postal mail and/or e-mail, analyze data, provide marketing research and assistance, process credit card payments and provide customer service. These contractors and partners are restricted from using information that identifies you (excluding any information that is aggregated or otherwise reasonably de-identified) in any way other than to provide services for us.'
+        },
+        {
+          title: 'Business Transfer.',
+          content: 'We may disclose, transfer, or assign to our affiliates or to one or more third parties the information collected about you as part of a merger, acquisition or other sale or transfer of any of the assets or business of RYDDO. In the unlikely event of our bankruptcy, insolvency, reorganization, receivership, or assignment for the benefit of creditors, or the application of laws or equitable principles affecting creditors\' rights generally, we may transfer your data to a successor or to a third party that purchases our assets arising from such circumstances. In such transactions, customer information generally is one of the transferred business assets but remains subject to the promises made in any pre-existing Privacy Policy.'
+        },
+        {
+          title: 'Protection of RYDDO and Others.',
+          content: 'We may disclose information about you to the government or to other third parties to comply with the law, applicable regulations, governmental and quasi-governmental requests, court orders, subpoenas and other legal process. We may also disclose information to third parties if needed to (i) enforce any of the terms of use for our Services or any investigation of potential violations thereof, (ii) detect, prevent, or otherwise address fraud, security or technical issues, (iii) protect against harm to our Services and (iv) protect our rights, property or safety or the rights, property or safety of our users or others.'
+        }
+      ]
+    },
+    {
+      id: 'information-from-children',
+      title: 'INFORMATION FROM CHILDREN',
+      content: '',
+      description: 'The Services are not directed at children and we have no intention of collecting information from young people. If we become aware that information from a child under 16 has been collected without the consent of such child\'s parent or guardian, we will use all reasonable efforts to delete such information.',
+      subsections: []
+    },
+    {
+      id: 'third-party-websites',
+      title: 'THIRD PARTY WEBSITES',
+      content: '',
+      description: 'As a convenience to our visitors, our Services may contain links to several websites that we believe offer useful information. The policies and procedures we describe here do not apply to those websites. We suggest contacting those websites directly for information on their privacy policies. RYDDO is not responsible for the privacy practices of third parties, regardless of whether they are linked or otherwise connected to our Services.',
+      subsections: []
+    },
+    {
+      id: 'cookies-and-interest-based-advertising',
+      title: 'COOKIES AND INTEREST-BASED ADVERTISING',
+      content: '',
+      description: 'Tracking tools we use. We and our service providers may collect information about users over time and across different websites and devices through the use of tracking tools including cookies, web beacons, device identifiers and other online information-gathering tools (collectively, "cookies"). While the cookies that we use may change from time to time, we use cookies for the same purposes, including:',
+      subsections: [
+        {
+          title: 'To authenticate your account and store your password if you are a registered account holder;',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'To monitor performance of our Services;',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'To learn more about the way you interact with our content;',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'To personalize your experience on our Sites and Social Media Services;',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'To optimize and tailor our Sites;',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'To make product recommendations and provide you with advertising content on our Services and elsewhere in which we think you will be interested;',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'To measure the effectiveness of personalized ads;',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'To pursue analytics and research to help us improve our Services and product offerings;',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'To better understand our Site visitors and their respective interests, including by learning about your online activities across devices.',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'Interest based advertising',
+          content: '',
+          type: 'paragraph',
+          paragraphs: [
+            'We and our service providers may use cookies to learn about what ads you see, what ads you click, and other actions you take on our Sites and other sites. This allows us to provide you with more useful and relevant ads on both RYDDO-owned and operated sites and unaffiliated sites. Interest-based ads, also sometimes referred to as personalized or targeted ads, are displayed to you based on information from activities such as purchasing on our Sites, visiting sites that contain our content or ads, interacting with our tools or using our payment services. In providing interest-based ads, we follow the Self-Regulatory Principles for Online Behavioral Advertising developed by the Digital Advertising Alliance (DAA) (a coalition of marketing, online advertising, and consumer advocacy organizations). See "Advertising Preferences and Opt-out" below to opt-out of such advertising.',
+            'Third party advertising management partners may help us display tailored content to our customers\' interests and serve interest-based advertising on our behalf. We do not provide any personal information to advertisers or to third party sites that display our interest-based ads. However, advertisers and other third-parties (including the ad networks, ad-serving companies, and other service providers they may use) may collect information through cookies about your browsing activity on our website to display interest based ads. Also, some third-parties may provide us information about you (such as the sites where you have been shown ads or demographic information) from offline and online sources that we may use to provide you more relevant and useful advertising.',
+            'Advertisers or ad companies may be working on their behalf sometimes use technology to serve the ads that appear on our sites directly to your browser. They automatically receive your IP address when this happens. They may also use cookies to measure the effectiveness of their ads and to personalize ad content. We do not have access to or control over cookies or other features that advertisers and third-party websites may use, and the information practices of these advertisers and third-party websites are not covered by our Privacy Policy. Please contact them directly for more information about their privacy practices.'
+          ]
+        },
+      ]
+    },
+    {
+      id: 'advertising-preferences-and-opt-out',
+      title: 'ADVERTISING PREFERENCES AND OPT-OUT',
+      content: '',
+      subsections: [
+        {
+          title: 'Cookies:',
+          content: 'The Help feature on most browsers will tell you how to prevent your browser from accepting new cookies, how to have the browser notify you when you receive a new cookie, or how to disable cookies altogether. If you set your browser to limit or remove cookies or otherwise block our use of cookies, some Site features may be unavailable or unable to function properly.',
+          type: 'paragraph'
+        },
+        {
+          title: 'Email marketing opt-out:',
+          content: 'If you wish to opt-out of any marketing email communications you may contact our customer service representatives at the phone number or email address below, under Questions. You also may follow opt-out unsubscribe instructions included in our promotional e-mail. Please note that removal of user information from our database or "opting-out" will not stop marketing or advertising content that is generated for distribution or is otherwise queued for transmission to you prior to the time when we can implement your request. Following receipt of a request from you, we will take reasonable steps to delete your information. If you opt-out of marketing communications, you will continue to receive transactional communications, such as e-mail or text notifications about your order status, recall information and other administrative information. Your information will remain in our database, but will be marked as inactive so you do not receive marketing materials from us.',
+          type: 'paragraph'
+        },
+        {
+          title: 'Facebook opt-out:',
+          content: '',
+          type: 'paragraph',
+          contentWithLinks: [
+            {
+              type: 'text',
+              content: 'Facebook may provide you with the ability to opt-out of certain advertising practices. See '
+            },
+            {
+              type: 'link',
+              link: {
+                text: 'https://www.facebook.com/help/1075880512458213',
+                href: 'https://www.facebook.com/help/1075880512458213',
+                className: 'text-pink-600 hover:text-pink-700 underline'
+              }
+            },
+            {
+              type: 'text',
+              content: ' However, we are unable to control Facebook\'s practices or disclosures. To opt-out of data collection by other Social Media Services, you should review that provider\'s privacy practices.'
+            }
+          ]
+        },
+        {
+          title: 'Do-Not-Track:',
+          content: 'Some browsers have a "do not track" feature that lets you tell websites that you do not want to have your online activities tracked. Our online Services do not support "do not track" preferences that may be available in your browser and are not capable of satisfying the preference you set.',
+          type: 'paragraph'
+        }
+      ]
+    },
+    {
+      id: 'data-security',
+      title: 'DATA SECURITY',
+      content: '',
+      description: 'We implement reasonable administrative, technical, and physical safeguards designed to protect the information we collect. However, no information system can be 100% secure, so we cannot guarantee the absolute security of your information.',
+      subsections: []
+    },
+    {
+      id: 'data-retention',
+      title: 'DATA RETENTION',
+      content: '',
+      description: 'We may retain your data for as long as we believe there is a business, legal or security reason for doing so, even if you are no longer an active customer.',
+      subsections: []
+    },
+    {
+      id: 'how-to-update-information',
+      title: 'HOW TO UPDATE INFORMATION',
+      content: '',
+      description: 'You may correct or update information collected about you through our Services by contacting us at the email address noted at the end of this Privacy Policy below, or by editing your information and preferences on the "Your Account" and "Profile and Password" pages. Following receipt of a request from you, we will take reasonable steps to update or correct your information.',
+      subsections: []
+    },
+    {
+      id: 'california-residents',
+      title: 'CALIFORNIA RESIDENTS',
+      content: '',
+      subsections: [
+        {
+          title: '',
+          content: '',
+          type: 'paragraph',
+          contentWithLinks: [
+            {
+              type: 'text',
+              content: 'Section 1798.83 of the California Civil Code permits California residents to request from a business, with whom the California resident has an established business relationship, certain information about the types of personal information the business has shared with third parties for those third parties\' direct marketing purposes and the names and addresses of the third parties with whom the business has shared such information during the immediately preceding calendar year. If you are a California resident, you may make one request each year by sending a request to ['
+            },
+            {
+              type: 'link',
+              link: {
+                text: 'info@Ryddo.com',
+                href: 'mailto:info@Ryddo.com',
+                className: 'text-pink-600 hover:text-pink-700 underline'
+              }
+            },
+            {
+              type: 'text',
+              content: '.]'
+            }
+          ]
+        }
+      ]
+    },
+    {
+      id: 'governing-law',
+      title: 'GOVERNING LAW',
+      content: '',
+      description: 'By choosing to visit our Services, you agree that any dispute over privacy or the terms contained in this Privacy Policy will be governed by the law of State of California, without reference to the choice of law or conflicts of law principles thereof, and will be subject to the dispute resolution clause contained in the terms of use applicable to the Services. You also agree to abide by any limitation on damages contained in the terms of use of our Services.',
+      subsections: []
+    },
+    {
+      id: 'changes-to-this-policy',
+      title: 'CHANGES TO THIS POLICY',
+      content: '',
+      description: 'We reserve the right, at our discretion, to modify our privacy practices and update and make changes to this Privacy Policy at any time. If we make any changes, we will change the Last updated date above and post the new Privacy Policy here. You should consult this Privacy Policy regularly for any changes.',
+      subsections: []
+    },
+    {
+      id: 'questions',
+      title: 'QUESTIONS',
+      content: '',
+      description: 'If there are any questions regarding this Privacy Policy, please contact us!',
+      subsections: []
+    }
+  ]
+};
