+++ core/lib/data/terms-conditions/shipping-returns.ts
@@ -1,0 +1,122 @@
+import type { TermsSection } from './interfaces';
+
+// Shipping, Returns & Exchanges page content
+export const shippingReturnsContent: TermsSection = {
+  id: 'shipping-returns',
+  title: 'Shipping, Returns, & Exchanges',
+  pageTitle: 'Shipping, Returns & Exchanges',
+  description: 'We want you to be completely satisfied with your purchase. Please review our shipping, return, and exchange policies below.',
+  accordionItems: [
+    {
+      id: 'shipping-policy',
+      title: 'Shipping Policy',
+      content: '',
+      subsections: [
+        {
+          title: '',
+          content: 'We make every effort to ship within 24-48 business hours once the purchases have received final approval. We do our best to ship products the same day, when possible. Orders that arrive before 12pm PST are often shipped the same day, when inventory is stocked at our retail location. Some products may be drop-shipped from third parties. Most of our electric scooters and bikes are shipped through UPS or Fedex ground services and delivery times range from 2 days in California to 4-5 days on the east coast but may vary greatly with delays. Call for a more definitive delivery estimation. We do offer express delivery through UPS, Fedex and USPS. We do not ship outside of the US, and can not ship battery operated products to Alaska and Hawaii at this time.',
+          type: 'paragraph'
+        }
+      ]
+    },
+    {
+      id: 'returns-exchanges-policy',
+      title: 'Returns & Exchanges Policy',
+      content: '',
+      description: 'We want you to be fully satisfied with every item that you purchase from Ryddo.com. However, if you need to return an item, we\'re here to help.',
+      subsections: [
+        {
+          title: '7-Day Test Ride',
+          titleStyle: 'normal',
+          content: 'Applies to select e-scooters, e-bikes or e-boards purchased from Ryddo.com. When the product includes the <strong>"7-Day Test Ride"</strong> banner on the product page you may return the product within (7) seven calendar days from the date of delivery, and we will refund the purchase price, <strong> minus(10%) of the full, non-discounted/sale, listed price,  on the product page and shipping costs both ways, as well as any mileage fees and repair costs, if applicable. The customer may also elect for a store credit for the same amount (Good for 12 months from the original purchase date).</strong> Subject to the following terms and conditions:',
+          type: 'paragraph'
+        },
+        {
+          title: 'Normal wear and tear is accepted. Acceptable wear and tear includes a small amount of tread wear equal to no more than <strong>150 miles of normal use</strong>, under typical usage patterns, as well as a minimal amount of wear on the breaks, and deck.',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'The product <strong>must be shipped by the end of the seventh day</strong> after delivery.',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'If the product is returned with more than 50 miles, you will be charged a $1.00 / mile fee for each mile over 50.',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'If the item has any damage beyond normal wear and tear, you will be billed for the cost to repair the damage at the rate of ($60/hr) in addition to the cost of any parts that are required to complete the repair and return the item to a "like new" condition.',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'A prepaid return label will be provided to you, upon request. When returning your item, please be sure to use the prepaid shipping label, and send all returned items in their UNDAMAGED ORIGINAL PACKAGING. If the original product packaging is damaged, we may assess a fee up to $50. We recommend that the customer take photos of the box prior to shipping. If original items that were included with the product are not included with the returned item, the cost of such items will be deducted from the amount refunded, at the cost of replacement to our company.',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: '30-Day Return Period for Electric Scooters & Bikes',
+          titleStyle: 'normal',
+          content: 'Any item purchased from Ryddo.com may be returned within (30) thirty calendar days from the date delivery of the item to you was completed, if the item remains <strong>new, unused, unopened, in its original packaging complete with all originally included materials and all tags still attached to the items. An RMA must be requested and approved or any returned will not be processed.</strong> Upon receipt, processing, and inspection of the returned item, we will refund the purchase price, minus discounts and credits, <strong>minus all shipping costs, outbound and return,</strong>  to you subject to the following terms and conditions:</strong>',
+          type: 'paragraph'
+        },
+        {
+          title: 'If you have removed the tags, used the products, or have destroyed the original packaging, we might be unable to accept the return. We reserve the right to refuse your return or exchange request, and the right to charge a restocking fee and other handling costs based on our sole and exclusive discretion.',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'Ryddo may charge additional fees for any accepted product returns which are found to be damaged upon receipt and inspection by Ryddo.',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'A prepaid return label will supplied upon request.  Please use this label and contact the shipping company for a pickup or drop-off. If insurance is included in the shipping it is the responsibility of the customer to obtain the required signatures to validate the insurance policy. UPS requires that a UPS driver, or an official UPS Customer Center must sign the documents. A UPS Store is not acceptable. The customer can use their shipping company of choice but Ryddo will not be responsible for lost or damaged products. It is at the customers\' discretion to purchase insurance for the package.',
+          content: '',
+          type: 'bullet'
+        },
+        {
+          title: 'Should the customer believe any items to be defective the customer must bring this to the attention of Ryddo prior to use. Once the product has been used it can no longer be returned for a refund and defective items would be covered under manufacturer warranty.',
+          content: '',
+          type: 'bullet'
+        },
+      ]
+    },
+    {
+      id: 'important',
+      title: '**IMPORTANT**',
+      content: '',
+      subsections: [
+        {
+          title: '',
+          content: 'Ryddo takes no responsibility for returned items that do not reach us. We strongly suggest that the customer insures any shipped packages for the minimum cost of the products.',
+          type: 'paragraph'
+        }
+      ]
+    },
+    {
+      id: 'frequently-asked-questions',
+      title: 'Frequently Asked Questions',
+      content: '',
+      subsections: [
+        {
+          title: 'Q. Can I make a return purchased in store?',
+          content: 'Your Ryddo experience is important to us, and we are therefore happy to provide exchanges or refunds for products purchased from either Ryddo.com or a Ryddo retail store, subject to the above return and exchange policy. Items purchased in, and returned to, a retail store location in the original unopened packaging, are entitled to a full refund or exchange within (30) calendar days of purchase.',
+          type: 'paragraph'
+        },
+        {
+          title: 'Q. What if I think my item might be faulty?',
+          content: 'We only ever want to provide the best quality products to our customers. If you believe that you have received a faulty item, please contact us or submit a request for return or exchange, along with any relevant images, within 24 hours of receipt on used products only. All claims will be covered under the manufacturer\'s warranty after the first 24 hours have lapsed. We aim to respond to all requests within 48-72 hours.',
+          type: 'paragraph'
+        },
+        {
+          title: 'Q. When will my purchase be refunded?',
+          content: 'We will process your refund as soon as we\'ve received and inspected your return. You will receive an email confirming that your refund has been processed, and the funds will appear in the original account used to make the purchase within ten (10) business days of the refund being issued, depending on your bank.',
+          type: 'paragraph'
+        }
+      ]
+    }
+  ]
+};
