+++ core/lib/data/terms-conditions/terms-of-use.ts
@@ -1,0 +1,136 @@
+import { COMPANY_INFO } from '~/lib/config/company-info';
+import type { TermsSection } from './interfaces';
+
+// Terms of Use page content
+export const termsOfUseContent: TermsSection = {
+  id: 'terms-of-use',
+  title: 'Terms of Use',
+  description: `Welcome to Ryddo.com, a website operated by Ryddo LLC. Please read the following Terms and Conditions of Use ("Terms of Use") carefully and ensure you understand them fully. These Terms of Use, including all documents referenced herein, govern your use of the Ryddo website (the "Services") and represent the entire understanding and agreement ("The Agreement") between RYDDO LLC and/or its parent(s), subsidiary(ies) and affiliate(s) (collectively "Ryddo", or "Ryddo.com") and you regarding your use of this website and supersede any prior statements or representations.<br /><br />BY USING OUR SERVICES, YOU ARE ACCEPTING THE PRACTICES DESCRIBED IN THESE TERMS OF USE. IF YOU DO NOT AGREE TO THESE TERMS OF USE, PLEASE DO NOT USE THE SERVICES. WE RESERVE THE RIGHT TO MODIFY OR AMEND THESE TERMS OF USE FROM TIME TO TIME WITHOUT NOTICE, BUT WILL NOTIFY YOU OF ANY MATERIAL CHANGES. YOUR CONTINUED USE OF OUR SERVICES FOLLOWING THE POSTING OF OR NOTICE OF CHANGES TO THESE TERMS WILL MEAN YOU ACCEPT THOSE CHANGES.`,
+  accordionItems: [
+    {
+      id: 'permission-to-use-content',
+      title: 'Permission to Use Content',
+      content: `Ryddo.com grants you a limited license to access and make personal, noncommercial use of this site and not to download (other than page caching) or modify it, in whole or in part, except with express written consent of Ryddo.com or where expressly permitted. You agree not to use or resell Ryddo.com or the Services it provides for any commercial purpose, including but not limited to sending unsolicited commercial messages. You agree not to use any services provided by Ryddo.com, if applicable, to send SPAM or to send messages to mailing lists for which you do not have the legal right to send or initiate the sending of such message to each recipient. In order to purchase products or services through Ryddo.com, you agree that you are 18 years of age or older, or you are 16 years of age or more and have the specific permission of a parent or legal guardian.`
+    },
+    {
+      id: 'site-services',
+      title: 'Site Services',
+      content: `Ryddo.com may change, suspend or discontinue any aspect of the products and services at any time, including the availability of any feature or content. Ryddo.com may also impose limits on certain features and services or restrict your access to parts or all of the product and services without notice or liability.`
+    },
+    {
+      id: 'copyright-and-trademark',
+      title: 'Copyright and Trademark',
+      content: `All content included on this site, photographs, images, moving images, sound, and illustrations ("Content"), is the property of RYDDO, or its Content suppliers or licensors and is protected by the United States and international copyrights, patents, trade secrets, or other forms proprietary rights. The compilation of all Content on this site is the exclusive property of RYDDO and is protected by the U.S. and international copyright laws. You may not modify, publish, transmit, participate in the transfer or sale of, reproduce, create derivative works from, distribute, perform, display, incorporate into another website, or in any other way exploit any of the content, in whole or in part without the specific permission of RYDDO. This prohibition includes but is not limited to, the use of any proprietary RYDDO design on any website, in emails, or in other media without express written permission from Ryddo.com.
+
+All trademarks, service marks, and trade names of Ryddo used or is using herein (collectively "Marks") are trademarks or registered trademarks of Ryddo or its affiliates, partners, vendors, or licensors. You may not use, copy, reproduce, republish, upload, post, transmit, distribute, or modify Marks in any way, including in advertising or publicity pertaining to distribution of materials on the Services, without Ryddo's prior written consent. You shall not use Ryddo's name or any language, pictures, or symbols which could based on Ryddo's judgment, imply our endorsement in any written or oral advertising or presentation, or brochure, newsletter, book, or other written material of whatever nature, without prior written consent.`
+    },
+    {
+      id: 'your-account-and-conduct',
+      title: 'Your Account and Member Conduct',
+      content: `To access many features on the site, and to purchase any goods or services, you may be required to create an account with Ryddo.com. Any personal information you provide in connection with your account will be covered by our Privacy Policy. You are responsible for the accuracy of the information included in your account, including updating your information as necessary, maintaining the confidentiality of your password. All activity conducted in connection with your account will be your responsibility, as you are deemed to be in sole possession and control of the confidential password necessary to access your account. You must immediately notify Ryddo.com of any unauthorized access or tampering of your account, or suspected breach of security.
+
+You also acknowledge and agree that Ryddo.com may preserve content and may also disclose content if required to do so by law or in the good faith belief that such preservation or disclosure is reasonably necessary to: (a) comply with legal process; (b) enforce these terms of service; (c) respond to claims that any content violates laws or the rights of third parties; or (d) protect the rights, property or personal safety of Ryddo.com, its users and the public.
+
+If in Ryddo.com's sole determination, you violate Ryddo.com's Member Conduct policy, Ryddo.com, may block you, and those using any email account associated with you from using any or all of the services available on Ryddo.com.
+
+RYDDO.COM RESERVES THE RIGHT IN ITS SOLE DISCRETION TO DISCLOSE TO THIRD PARTIES, INCLUDING PRODUCT RECIPIENTS, THEIR COUNSEL, OR LAW ENFORCEMENT AUTHORITIES, RELEVANT INFORMATION ABOUT A COMMUNICATION, INCLUDING, BUT NOT LIMITED TO, THE SENDER NAME, ACCOUNT INFORMATION, MEMBER AND COOKIE NUMBERS, IP ADDRESSES, TRANSMISSION DATA, PAST AND SUBSEQUENT SERVICE USE, AND ANY OTHER INFORMATION DEEMED NECESSARY BY RYDDO.COM TO ASSIST THIRD PARTIES, THEIR COUNSEL AND LAW ENFORCEMENT INVESTIGATE VIOLATIONS OF THE MEMBER CONDUCT POLICY. THESE DISCLOSURES MAY BE MADE WITH OR WITHOUT PRIOR NOTICE TO YOU. BY USING THE SERVICE, YOU IRREVOCABLY CONSENT TO SUCH DISCLOSURES.`
+    },
+    {
+      id: 'deletion-or-deactivation',
+      title: 'Deletion or Deactivation of Your Content and/or Your Account',
+      content: `You may terminate this Agreement at any time (i) by sending notice in writing to Ryddo.com at 334 S Main St #5017, Los Angeles, CA. 90013 confirming such termination, (ii) by removing all Your Content from your account, (iii) by deleting your account and thereafter by ceasing to use the Website.`
+    },
+    {
+      id: 'payment-and-purchases',
+      title: 'Payment, Purchase of Products or Services',
+      content: `You may be required to provide Ryddo with a valid credit card, debit card, or another payment account ("Payment Method") in order to purchase any products or services provided by Ryddo. When you add a Payment Method to your account, you will be asked to provide customary billing information. You must provide accurate, current, and complete information when adding a Payment Method. You agree to pay all fees and charges incurred through your purchase of any products or services at the time of purchase and at the rates in effect for the billing period in which such fees and charges are incurred. RYDDO does not provide refunds or credits if the price for particular Products or Services previously purchased by you is lowered or is part of a promotional offer. All Products sold through Ryddo.com are subject to our Return and Exchange Policy. For more information, <strong><em>review our full Return and Exchange Policy</em></strong>.`
+    },
+    {
+      id: 'limitation-of-liability',
+      title: 'Limitation of Liability',
+      content: `UNDER NO CIRCUMSTANCES, INCLUDING BUT NOT LIMITED TO ITS OWN NEGLIGENCE, SHALL RYDDO.COM BE LIABLE FOR ANY DIRECT, INDIRECT, CONSEQUENTIAL, SPECIAL OR EXEMPLARY DAMAGES THAT RESULT FROM THE USE OF, OR THE INABILITY TO USE, THE MATERIALS IN THIS SITE, EVEN IF RYDDO.COM OR A RYDDO.COM AUTHORIZED REPRESENTATIVE HAS BEEN ADVISED OF THE POSSIBILITY OF SUCH DAMAGES. APPLICABLE LAW MAY NOT ALLOW THE LIMITATION OR EXCLUSION OF LIABILITY OR INCIDENTAL OR CONSEQUENTIAL DAMAGES, SO THE ABOVE LIMITATION OR EXCLUSION MAY NOT APPLY TO YOU. IN NO EVENT SHALL RYDDO.COM'S TOTAL LIABILITY TO YOU FOR ALL DAMAGES, LOSSES AND CAUSES OF ACTION (WHETHER IN CONTRACT, TORT INCLUDING BUT NOT LIMITED TO NEGLIGENCE OR OTHERWISE) EXCEED THE AMOUNT PAID BY YOU, IF ANY, FOR ACCESSING THIS SITE.`
+    },
+    {
+      id: 'disclaimer-of-warranties',
+      title: 'Disclaimer of Warranties',
+      content: `The materials in this site are provided "as is" and without warranties of any kind, either expressed or implied. To the fullest extent permissible pursuant to applicable law, Ryddo.com disclaims all warranties, express or implied, including but not limited to implied warranties of merchantability and fitness for a particular purpose and non-infringement. Ryddo.com does not warrant that the functions contained in the materials will be uninterrupted or error-free, that defects will be corrected or that this site or the server that makes it available are free of viruses or other harmful components. Ryddo.com does not warrant or make any representation regarding the use or the results of the use of the materials in this site in terms of their correctness, accuracy, reliability or otherwise. You (and not Ryddo.com) assume the entire cost of all necessary servicing, repair or correction. Applicable law may not allow the exclusion of implied warranties so the above exclusion may not apply to you.`
+    },
+    {
+      id: 'product-information-disclaimer',
+      title: 'Product Information Disclaimer',
+      content: `It is our intent to provide the most accurate and up-to-date information available throughout the site and in our communications with you. Occasionally, information on Ryddo.com may contain typographical errors, inaccuracies or omissions that may relate to product descriptions, pricing and/or availability. Ryddo.com reserves the right to correct any errors, inaccuracies or omissions and to change or update information at any time without prior notice, including after your order has been submitted. In addition, we reserve the right to cancel orders due to lack of availability, including after your order has been submitted. We apologize for any inconvenience this may cause you.`
+    },
+    {
+      id: 'user-submissions',
+      title: 'User Submissions',
+      content: `On certain sections of this Site, users may be permitted to: post, display and/or publish ("post") comments about Ryddo's products or other materials ("User Content"). You acknowledge and agree that you are solely responsible for any User Content that you post. You further acknowledge and agree that Ryddo has no responsibility for, and makes no representations concerning, any User Content posted at this website and will not be liable for any User Content. With respect to all User Content that you post, you represent and warrant that: (i) the material is fully original to you; (ii) such User Content will not infringe or violate the rights of any person or entity, or violate any governmental rule, regulation statute or law, or the Terms if Use; (iii) no money will be owing to any person or entity as a result of the posting of the User Content or its use as contemplated by these Terms if Use; and (iv) you will be responsible for all User Content submitted through your account, and for all purposes under these Terms and Conditions, all User Content submitted from your account shall be deemed to have been submitted by you.`
+    },
+    {
+      id: 'user-content-license',
+      title: 'User Content License',
+      content: `User Submissions remain the intellectual property of the individual user. By submitting content to Ryddo, you expressly grant Ryddo a non-exclusive, perpetual, irrevocable, royalty-free, fully paid-up worldwide, fully sub-licensable right to use, reproduce, modify, adapt, publish, translate, create derivative works from, distribute, transmit, perform and display such content, and irrevocably waive any and all "moral rights" (howsoever known) you may have in such User Content and/or consent to Ryddo (or any of its nominees, licensees or successors) engaging in conduct which may otherwise infringe any "moral rights" in your User Content. Further, you understand and agree that Ryddo may retain, but not display, distribute or perform, server copies of User Content that has been deleted or removed. You understand that Ryddo performs technical functions necessary to offer its website and its Services, including but not limited to transcoding and/or reformatting content to allow its use through the Website. You agree that you will make no commercial use of any User Content obtained through this website, without the express written agreement of Ryddo.`
+    },
+    {
+      id: 'innapropriate-user-submissions',
+      title: 'Inappropriate User Submissions',
+      content: `Ryddo does not encourage, and does not welcome User Submissions that result from any activity that: (i) may create a risk of harm, loss, physical or mental injury, emotional distress, death, disability, disfigurement, or physical or mental illness to you, to any other person, or to any animal; (ii) may create a risk of any other loss or damage to any person or property; or (iii) may constitute a crime or tort.  You agree that you have not and will not engage in any of the foregoing activities in connection with producing your User Submission. Without limiting the foregoing, you agree that in conjunction with your submission, you will not inflict emotional distress on other people, will not humiliate other people (publicly or otherwise), will not assault or threaten other people, will not enter onto private property without permission, will not impersonate any other person or misrepresent your affiliation, title, or authority, and will not otherwise engage in any activity that may result in injury, death, property damage, and/or liability of any kind. Ryddo will reject any User Submissions in which Ryddo believes, in its sole discretion, that any such activities have occurred.  If notified by a user of a submission that allegedly violates any provision of these Terms of Use, Ryddo reserves the right to determine, in its sole discretion, if such a violation has occurred, and to remove any such submission from the Services at any time and without notice.`
+    },
+    {
+      id: 'termination',
+      title: 'Termination',
+      content: `Either you or Ryddo may suspend or terminate your right to use of Ryddo's Services at any time, for any reason or for no reason. Ryddo may also block your access to our Services in the event that (a) you breach these Terms of Use; (b) we are unable to verify or authenticate any information you provide to us; or (c) we believe that your actions may cause financial loss or legal liability for you, our users or us.`
+    },
+    {
+      id: 'disputes-and-jurisdiction',
+      title: 'Disputes and Jurisdiction',
+      content: `In the event you violate these Terms of Use, in RYDDO's sole discretion, Ryddo.com may at any time, terminate your account, suspend the delivery of or access to, services, refuse membership and/or terminate your membership or subscription, in addition to all other legal rights and remedies available to it.
+
+This Agreement shall be governed by and construed in accordance with the laws of the State of California, without giving effect to any principles of conflicts of law. Any dispute relating in any way to your visit to Ryddo.com or to products you purchase through Ryddo.com shall be submitted to confidential arbitration in Los Angeles, California, except that, to the extent you have in any manner violated or threatened to violate Ryddo.com's intellectual property rights, Ryddo.com may seek injunctive or other appropriate relief in any state or federal court in the State of California, and you consent to exclusive jurisdiction and venue in such courts. Arbitration under this Agreement shall be conducted under the rules then prevailing of the American Arbitration Association. The arbitrator's award shall be binding and may be entered as a judgment in any court of competent jurisdiction. To the fullest extent permitted by applicable law, no arbitration under this Agreement shall be joined to an arbitration involving any other party subject to this Agreement, whether through class arbitration proceedings or otherwise, and you agree that you will not participate as a class member or class representative in any action against Ryddo.com.`
+    },
+    {
+      id: 'miscellaneous',
+      title: 'Miscellaneous',
+      content: `If any provision of this Agreement shall be unlawful, void, or for any reason is unenforceable, then that provision shall be deemed severable from this Agreement and shall not affect the validity and enforceability of any remaining provisions. This is the entire agreement between the parties relating to the subject matter herein and shall not be modified except in writing, signed by both parties.
+
+Ryddo.com makes no representation that materials in the site are appropriate or available for use outside the United States. Those who choose to access this site from outside the United States do so on their own initiative and are responsible for compliance with local laws, if and to the extent local laws are applicable. Individuals also agree to comply with all applicable United States export controls.`
+    },
+    {
+      id: 'severability',
+      title: 'Severability',
+      content: `Should one or more provisions of these Terms of Use be found to be unlawful, void or unenforceable, such provision(s) shall be deemed severable and will not affect the validity and/or enforceability of the remaining provisions of the Terms of Use, which will remain in full force and effect.`
+    },
+    {
+      id: 'entire-agreement',
+      title: 'Entire Agreement',
+      content: `These Terms of Use, together with the Privacy Policy constitute the entire Agreement between you and Ryddo.com with respect to your use of the Website and supersede any prior agreement between you and Ryddo.com. Any modifications to this Agreement must be made in writing.`
+    },
+    {
+      id: 'changes-to-terms-of-use',
+      title: 'Changes to Terms of Use',
+      content: `We reserve the right to change, alter, replace or otherwise modify these Terms of Use at any time. The date of the last modification is stated at the end of these Terms of Use. It is your responsibility to check this page from time to time for updates. Your continued use of the Website will be deemed as irrevocable acceptance of any changes and revisions to these Terms of Use.`
+    },
+    {
+      id: 'copyright-agent',
+      title: 'Copyright Agent',
+      content: `Ryddo.com respects the intellectual property of others. If you believe that your work has been copied in a way that constitutes copyright infringement, or your intellectual property rights have been otherwise violated, please contact Ryddo.com's Copyright Agent and provide the following information:<br />
+<br />
+<ul>
+<li>An electronic or physical signature of the person authorized to act on behalf of the owner of the copyright or other intellectual property interest;</li>
+<li>A description of the copyrighted work or other intellectual property that you claim has been infringed;</li>
+<li>A description of where the material that you claim is infringing is located on the site;</li>
+<li>Your address, telephone number, and email address;</li>
+<li>A statement by you that you have a good faith belief that the disputed use is not authorized by the copyright owner, its agent, or the law;</li>
+<li>A statement by you, made under penalty of perjury, that the above information in your Notice is accurate and that you are the copyright or intellectual property owner or authorized to act on the copyright or intellectual property owner's behalf.</li>
+</ul>
+<br />
+Please send your copyright information to:<br /><br />
+
+${COMPANY_INFO.copyrightAgent.company}<br />
+${COMPANY_INFO.copyrightAgent.address}<br />
+${COMPANY_INFO.copyrightAgent.city}, ${COMPANY_INFO.copyrightAgent.state}. ${COMPANY_INFO.copyrightAgent.zip}<br /><br />
+
+Attention: ${COMPANY_INFO.copyrightAgent.attention}<br />
+Email: ${COMPANY_INFO.copyrightAgent.email}`
+    }
+  ]
+};
